+++ database/migrations/2023_10_27_204640_bills_has_products.php
@@ -18,11 +18,11 @@ class BillsHasProducts extends Migration
             $table->unsignedBigInteger('id_bill');
             $table->unsignedBigInteger('id_product');
 
-            $table->decimal('price');
+            $table->bigInteger('price');
             $table->bigInteger('stock');
             $table->bigInteger('discount');
 
-            $table->decimal('total_prices');
+            $table->bigInteger('total_prices');
             $table->timestamps();
         });
     }
