+++ database/migrations/2023_11_08_210849_create_budgets_has_products_table.php
@@ -18,11 +18,11 @@ class CreateBudgetsHasProductsTable extends Migration
             $table->unsignedBigInteger('id_budget');
             $table->unsignedBigInteger('id_product');
 
-            $table->decimal('price');
+            $table->bigInteger('price');
             $table->bigInteger('stock');
             $table->bigInteger('discount');
 
-            $table->decimal('total_prices');
+            $table->bigInteger('total_prices');
             $table->timestamps();
         });
     }
